requirements.txt
discord.py
python-dotenv
pillow
requests
pyrender
trimesh
aiofiles
aioshutil
aiopathlib
cryptography
nsfw-detector